server/requirements.txt
openenv-core[core]>=0.2.1
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
gradio>=5.0.0,<7.0.0
pillow>=10.0.0
requests>=2.31.0
groq>=0.9.0
openai>=1.0.0
python-dotenv>=1.0.0